add createConta

+++ backend/msconta/src/main/java/br/net/dac/msconta/model/dto/ContaDTO.java
@@ -3,9 +3,9 @@ package br.net.dac.msconta.model.dto;
 import java.sql.Date;
 
 public class ContaDTO {
-    private long id;
-    private long idGerente;
-    private long idCliente;
+    private Long id;
+    private Long idGerente;
+    private Long idCliente;
     private String numeroConta;
     private Date dataCriacao;
     private double saldo;
@@ -14,7 +14,7 @@ public class ContaDTO {
     public ContaDTO() {
     }
 
-    public ContaDTO(int id, String numeroConta, Date dataCriacao, double saldo, double limite) {
+    public ContaDTO(long id, String numeroConta, Date dataCriacao, double saldo, double limite) {
         this.id = id;
         this.numeroConta = numeroConta;
         this.dataCriacao = dataCriacao;
@@ -24,14 +24,14 @@ public class ContaDTO {
 
     // Getters e Setters
 
-    public long getId() {return id;}
-    public void setId(long id) {this.id = id;}
+    public Long getId() {return id;}
+    public void setId(Long id) {this.id = id;}
 
-    public long getIdGerente() {return idGerente;}
-    public void setIdGerente(long idGerente) {this.idGerente = idGerente;}
+    public Long getIdGerente() {return idGerente;}
+    public void setIdGerente(Long idGerente) {this.idGerente = idGerente;}
 
-    public long getIdCliente() {return idCliente;}
-    public void setIdCliente(long idCliente) {this.idCliente = idCliente;}
+    public Long getIdCliente() {return idCliente;}
+    public void setIdCliente(Long idCliente) {this.idCliente = idCliente;}
 
     public String getNumeroConta() {return numeroConta;}
     public void setNumeroConta(String numeroConta) {this.numeroConta = numeroConta;}
@@ -39,9 +39,9 @@ public class ContaDTO {
     public Date getDataCriacao() {return dataCriacao;}
     public void setDataCriacao(Date dataCriacao) {this.dataCriacao = dataCriacao;}
 
-    public double getSaldo() {return saldo;}
-    public void setSaldo(double saldo) {this.saldo = saldo;}
+    public Double getSaldo() {return saldo;}
+    public void setSaldo(Double saldo) {this.saldo = saldo;}
 
-    public double getLimite() {return limite;}
-    public void setLimite(double limite) {this.limite = limite;}
+    public Double getLimite() {return limite;}
+    public void setLimite(Double limite) {this.limite = limite;}
 }
